inference/requirements.txt
tensorflow==2.15.0
numpy>=1.23.5,<2.0.0
opencv-python-headless>=4.8.0
pillow>=10.0.0
gradio==4.44.1
python-multipart==0.0.9
scikit-learn>=1.3.0
requests>=2.31.0

